+++ lib/dealog_backoffice/messages/projectors/deleted_message.ex
@@ -1,0 +1,24 @@
+defmodule DealogBackoffice.Messages.Projectors.DeletedMessage do
+  use Commanded.Projections.Ecto,
+    application: DealogBackoffice.App,
+    name: "Messages.Projectors.DeletedMessage",
+    consistency: :strong
+
+  alias DealogBackoffice.Messages.Events.MessageDeleted
+
+  alias DealogBackoffice.Messages.Projections.DeletedMessage
+
+  project(%MessageDeleted{} = deleted, metadata, fn multi ->
+    Ecto.Multi.insert(
+      multi,
+      :deleted_message,
+      %DeletedMessage{
+        id: deleted.message_id,
+        title: deleted.title,
+        body: deleted.body,
+        status: deleted.status,
+        inserted_at: metadata.created_at
+      }
+    )
+  end)
+end